Allow user to disable importing when use'ing Akari
WWW::Shorten::generic's custom import method did bad stuff so
we get rid of it and use Exporter as it was meant to be. Alas,
things would not be as easy as wanted, and a workaround had to
be added to a quirk of WWW::Shorten.

Also allows makeashorterlink and makealongerlink to be called
through the object interface.

Add tests for the new "Exporter-powered" :short imports.